Applied Preserve whole object to evaluation iterator constructors.
Since the construction of those evaluation iterators became a bit
hairy, and I need to implement more of those (at least the join),
I figured it would be worth to make the interface nicer.
Thus, I checked the interface and figured that tearing the
RelAlgExpr-Nodes apart in the evaluator is a smell and fixed this.

Detailed changes:
 - Applied Preserve Whole Object to ProjectionEvaluationIterator constructor
   - Introduced parameter: Projection p
   - Changed: selectedFields is obtained from whole object p
   - Removed Parameter: selectedFields
 - Applied Preserve Whole Object to ProductEvaluationIterator constructor
   - Introduced Parameter: Product p
   - Changed: left subexpression is obtained from whole object p
   - Changed: right subexpresson is obtained from whole object p
   - Removed Parameter: pLeft
   - Removed Parameter: pRight
 - Applied Preserve Whole Object to SelectionEvaluationIterator constructor
   - Introduced Parameter: Selection s
   - Changed: predicate is obtained from whole object s
   - Removed Parameter: predicate