Implemented the abstract tree reorderer.
The abstract tree reorderer encapsulates the mundane parts of finding
and interesting node to move downwards the tree and moving it downwards
(which are traversing and reconstructing the tree).
The tree reorderer was implemented as a state pattern after a certain
internal refactoring, which is omitted here, as no one saw the original
version and thus, no one might need to reconstruct those changes.

Detailed changes:
 - Classes added: TreeReorderer, TreeReordererState, SearchingAcceptedNodeState,
   MovingProjectionState, MovingSelectionstate
 - Interfaces added: IReorderingScheme, IOptimization
 - Added interfaces for tree reorderers and used parameters
 - Test added: Scheme accepts nothing results in trivial identity
 - Test added: Trivial moves keep the optimization trivial
 - Test added: Selection moves along projection
 - Test added: Projection moves along projections
 - Test added: LoadTable stops pushing down is red
 - Test added: RenameTable stops the pushing down
 - Test added: No Operation for products does nothing
 - Test added: Product children are transformed
 - Test added: Node moved left along product
 - Test added: Node moved right along product
 - Test added: Treereorderer can optimize multiple times